+++ xnode/xnode.h
@@ -33,7 +33,7 @@ along with elim.  If not, see <http://www.gnu.org/licenses/>.
 
 char    * xnode_unescape_html     ( const char *html );
 xmlnode * xnode_alist_item_string ( const char *name, const char *value );
-xmlnode * xnode_alist_item_integer( const char *name, int         value );
+xmlnode * xnode_alist_item_integer( const char *name, long        value );
 xmlnode * xnode_alist_item_boolean( const char *name, gboolean    value );
 xmlnode * xnode_alist_item_xnode  ( const char *name, xmlnode    *value );
 xmlnode * xnode_alist_item_enum   ( const char *name, int val, const char *t );
